tests: Add SKIP_TEST_* variables to allow these tests to be skipped.
By setting these variables, we can skip tests that fail on
RHEL 5.
(cherry picked from commit 145f35badfd9ca4a6d9f54f7732566a5e5114876)

+++ tests/regressions/rhbz690819.sh
@@ -35,6 +35,11 @@
 set -e
 export LANG=C
 
+if [ -n "$SKIP_TEST_RHBZ690819_SH" ]; then
+    echo "$0: test skipped because environment variable is set."
+    exit 0
+fi
+
 rm -f test.img
 
 ../../fish/guestfish sparse test.img 100M
